requirements.txt
SQLAlchemy==2.0.31
python-dotenv==1.0.1
asyncpg==0.29.0
fastapi==0.111.0
alembic==1.13.2
loguru==0.7.2
pydantic==2.8.2
fastapi_users==13.0.0
fastapi-users-db-sqlalchemy==6.0.1
uvicorn==0.30.1
gunicorn==22.0.0
google-api-core==2.19.1
google-api-python-client==2.136.0
google-auth==2.31.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.2

